added data to joke
