+++ dad-joke-client/src/components/Joke.js
@@ -1,0 +1,16 @@
+import React from 'react'
+
+const Joke = ({item}) => {
+
+  console.log(item)
+
+  return (
+    <>
+      <article>
+        <h3>{item.joke}</h3>
+      </article>
+    </>
+  )
+}
+
+export default Joke
