+++ dad-joke-client/src/components/axiosWithAuth.js
@@ -2,9 +2,8 @@ import axios from 'axios'
 
 const axiosWithAuth = () => {
     return axios.create({
-        baseURL: 'http://localhost:3300/api/jokes',
         headers: {
-            token: localStorage.getItem('token')
+            authorization: localStorage.getItem('token')
         }
     })
 }
